use a shaded cube for project_point2d too.
